move logic back down to departure card, more moving
