+++ apps/site/assets/ts/stop/components/StopPageDepartures.tsx
@@ -1,21 +1,12 @@
-import { filter, groupBy, reject, sortBy } from "lodash";
+import { filter, groupBy, sortBy } from "lodash";
 import React, { ReactElement, useState } from "react";
 import { Alert, Route } from "../../__v3api";
 import DeparturesFilters, { ModeChoice } from "./DeparturesFilters";
-import {
-  isRailReplacementBus,
-  isSubwayRoute,
-  modeForRoute
-} from "../../models/route";
+import { isRailReplacementBus, modeForRoute } from "../../models/route";
 import DepartureCard from "./DepartureCard";
 import { alertsByRoute, isInNextXDays } from "../../models/alert";
 import { DepartureInfo } from "../../models/departureInfo";
-import {
-  GroupedRoutePatterns,
-  RoutePatternWithPolyline,
-  sortedGroupedRoutePatterns
-} from "../../models/route-patterns";
-import { departureInfoInRoutePatterns } from "../../helpers/departureInfo";
+import { GroupedRoutePatterns } from "../../models/route-patterns";
 
 interface StopPageDeparturesProps {
   routes: Route[];
@@ -33,14 +24,6 @@ const modeSortFn = ({ type }: Route): number => {
     return 2;
   }
   return type;
-};
-
-const isNoncanonicalAndNoDepartures = (
-  routePatterns: RoutePatternWithPolyline[],
-  departures: DepartureInfo[]
-): boolean => {
-  const isNonCanonical = !routePatterns.find(rp => !!rp.canonical);
-  return isNonCanonical && departures.length === 0;
 };
 
 const StopPageDepartures = ({
@@ -70,40 +53,17 @@ const StopPageDepartures = ({
         />
       )}
       <ul className="stop-departures list-unstyled">
-        {sortBy(filteredRoutes, [modeSortFn, "sort_order"]).map(route => {
-          const groupedByHeadsign = groupedRoutePatterns[route.id];
-          let sortedRoutePatternsByHeadsign = sortedGroupedRoutePatterns(
-            groupedByHeadsign
-          );
-          const departuresForRoute = departureInfos.filter(
-            d => d.route.id === route.id
-          );
-          if (isSubwayRoute(route)) {
-            // remove certain headsigns based on route pattern and departures
-            sortedRoutePatternsByHeadsign = reject(
-              sortedRoutePatternsByHeadsign,
-              entry => {
-                const [, { route_patterns: routePatterns }] = entry;
-                const departures = departuresForRoute.filter(d =>
-                  departureInfoInRoutePatterns(d, routePatterns)
-                );
-                return isNoncanonicalAndNoDepartures(routePatterns, departures);
-              }
-            );
-            // don't render a route card if there's no headsigns left to show
-            if (sortedRoutePatternsByHeadsign.length === 0) return null;
-          }
-
-          return (
-            <DepartureCard
-              key={route.id}
-              route={route}
-              routePatternsByHeadsign={sortedRoutePatternsByHeadsign}
-              alertsForRoute={groupedAlerts[route.id] || []}
-              departuresForRoute={departuresForRoute}
-            />
-          );
-        })}
+        {sortBy(filteredRoutes, [modeSortFn, "sort_order"]).map(route => (
+          <DepartureCard
+            key={route.id}
+            route={route}
+            routePatternsByHeadsign={groupedRoutePatterns[route.id]}
+            alertsForRoute={groupedAlerts[route.id] || []}
+            departuresForRoute={departureInfos.filter(
+              d => d.route.id === route.id
+            )}
+          />
+        ))}
       </ul>
     </div>
   );
